chore: style announcements
