+++ package.json
@@ -26,6 +26,7 @@
     "browserify-zlib": "^0.2.0",
     "clsx": "^1.1.1",
     "crypto-browserify": "^3.12.0",
+    "date-fns": "^2.29.2",
     "material-ui-phone-number": "^2.2.6",
     "path-browserify": "^1.0.1",
     "react": "^18.2.0",
